added a manager
