--- Assets/Prefabs/GameObjects/Player.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9fd04531e64443bd48185824dfb99762
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
