--- Assets/Prefabs/GameObjects/PlayerCanvas.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e53eca77d7c8c568ba7b3696c32149a8
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
