+++ Assets/Prefabs/Player/Player.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 66cc9a91696df80618cac7cae0d2c65a
+guid: 47d012ba307b351fbb5f864e8e4fc4b9
 PrefabImporter:
   externalObjects: {}
   userData:
