--- Assets/Prefabs/PlayerSkins/AgentSmith.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 17f6f7648905e3937b7c7fe69f10f071
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
