--- Assets/Prefabs/PlayerSkins/Alien.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 15303f37197bdd795be007b4a6c87015
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
