--- Assets/Prefabs/PlayerSkins/DoctorSins.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6e2300e26e43ca8d2890a092e66dc426
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
