--- Assets/Prefabs/PlayerSkins/Neo.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 740d13ae44ec3b81a9487a476bcb46a2
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
