--- Assets/Prefabs/PlayerSkins/OrangeGirl.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 99cdc2d445d3f65f09658fa54a315c64
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
