--- Assets/Prefabs/PlayerSkins/Police.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c72ec07291379946ba8261f4a8af1eaa
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
